Propensity Score Weighting Using BART

This page explains the details of estimating weights from Bayesian additive regression trees (BART)-based propensity scores by setting method = "bart" in the call to weightit or weightitMSM. This method can be used with binary, multinomial, and continuous treatments.

In general, this method relies on estimating propensity scores using BART and then converting those propensity scores into weights using a formula that depends on the desired estimand. This method mainly relies on gbart from the BART package.

Binary Treatments

For binary treatments, this method estimates the propensity scores using gbart. The following estimands are allowed: ATE, ATT, ATC, ATO, and ATM. The weights for the ATE, ATT, and ATC are computed from the estimated propensity scores using the standard formulas, the weights for the ATO are computed as in Li & Li (2018), and the weights for the ATM (i.e., average treatment effect in the equivalent sample "pair-matched" with calipers) are computed as in Yoshida et al. (2017). Weights can also be computed using marginal mean weighting through stratification for the ATE, ATT, and ATC. See get_w_from_ps for details.

Multinomial Treatments

For multinomial treatments, this method estimates the propensity scores using mbart. The following estimands are allowed: ATE, ATT, ATO, and ATM. The weights for each estimand are computed using the standard formulas or those mentioned above. Weights can also be computed using marginal mean weighting through stratification for the ATE, ATT, and ATC. See get_w_from_ps for details.

Continuous Treatments

For continuous treatments, the generalized propensity score is estimated using gbart. In addition, kernel density estimation can be used instead of assuming a normal density for the numerator and denominator of the generalized propensity score by setting use.kernel = TRUE. Other arguments to density can be specified to refine the density estimation parameters. plot = TRUE can be specified to plot the density for the numerator and denominator, which can be helpful in diagnosing extreme weights.

Longitudinal Treatments

For longitudinal treatments, the weights are the product of the weights estimated at each time point.

Sampling Weights

Sampling weights are not supported.

Missing Data

In the presence of missing data, the following value(s) for missing are allowed:

"ind" (default)

First, for each variable with missingness, a new missingness indicator variable is created which takes the value 1 if the original covariate is NA and 0 otherwise. The missingness indicators are added to the model formula as main effects. The missing values in the covariates are then replaced with 0s. The weight estimation then proceeds with this new formula and set of covariates. The covariates output in the resulting weightit object will be the original covariates with the NAs.

"hot"

The function that fits the BART model (i.e., gbart or mbart) performs hot-deck imputation, which involves randomly selecting a value to impute the missing covariate with from another unit in the dataset with a nonmissing value for that variable. The BART authors recommend only doing this for small amounts of missingness. Doing so will slightly weaken the estimated relationship between the covariate with missingness and the outcome (what the authors describe as being "biased toward the null").

Additional Arguments

link

The link used in the BART model for the propensity scores with binary or multinomial treatments. The allowable options are "logit" and "probit". "probit" is much faster and is the default; there is generally little reason to choose "logit". This option is used to supply the argument to type in gbart and mbart.

use.mbart2

For multinomial treatments, whether to use mbart2 or mbart to fit the BART model for the propensity score. mbart uses a conditional probability specification to fit the models; first the probability of one category versus all the others is modeled, then the probability of the second category versus all the other is model using only those who are not in the first category, and so on. mbart2 uses a series of binary BART models for each category versus the others. The package authors recommend using mbart for "few" categories, so this is the default (i.e., use.mbart2 = FALSE). mbart tends to be faster.

mc.cores

The number of cores to enage when performing parallelization (not available on Windows). If greater than 1, the multicore version of the corresponding BART function will be used (e.g., mc.gbart instead of gbart and mc.mbart instead of mbart). Parallelization requires no extra effort or knowledge to perform and can dramatically speed up model fitting, so it is highly recommended. It's as simple as setting mc.cores = 4 in the call to weightit. This value is passed on to the mc.cores argument in the fitting function. The default is 1 for no parallelization. When using parallelization, an argument should be passed to seed to ensure the estimation is replicable; setting the seed outside the function will not always work. seed should probably be set differently each time weightit is called if replicability is not important because otherwise it will remain at its default value of 99 every time, which guarantees replicability but also dependence on that particular (arbitrary) value.

All arguments to the BART fitting functions can be passed through weightit or weightitMSM, with the following exceptions:

type and ntype are ignored because the type is determined by the form of the treatment variable and the argument supplied to link.

The ntree argument can optionally be supplied as n.trees to be consistent with method = "gbm".

For continuous treatments only, the following arguments may be supplied:

density

A function corresponding to the conditional density of the treatment. The standardized residuals of the treatment model will be fed through this function to produce the numerator and denominator of the generalized propensity score weights. If blank, dnorm is used as recommended by Robins et al. (2000). This can also be supplied as a string containing the name of the function to be called. If the string contains underscores, the call will be split by the underscores and the latter splits will be supplied as arguments to the second argument and beyond. For example, if density = "dt_2" is specified, the density used will be that of a t-distribution with 2 degrees of freedom. Using a t-distribution can be useful when extreme outcome values are observed (Naimi et al., 2014). Ignored if use.kernel = TRUE (described below).

use.kernel

If TRUE, uses kernel density estimation through the density function to estimate the numerator and denominator densities for the weights. If FALSE, the argument to the density parameter is used instead.

bw, adjust, kernel, n

If use.kernel = TRUE, the arguments to the density function. The defaults are the same as those in density except that n is 10 times the number of units in the sample.

plot

If use.kernel = TRUE, whether to plot the estimated density.

Additional Outputs

obj

When include.obj = TRUE, the BART fit(s) used to generate the predicted values. The specific function used to fit the model depends on the inputs.

Details

BART works by fitting a sum-of-trees model for the treatment or probability of treatment. The number of trees is determiend by the ntree argument. Bayesian priors are used for the hyperparameters, so the result is a posterior distribution of predicted values for each unit. The mean of these for each unit is taken as the (generalized) propensity score. Although the hyperparameters governing the priors can be modified by supplying arguments to weightit that are passed to the BART fitting function, the default values tend to work well and require little modification (though the defaults differ for continuous and categorical treatments; see the gbart documentation for details). Unlike many other machine learning methods, no loss function is optimized and the hyperparameters do not need to be tuned (e.g., using cross-validation). BART tends to balance sparseness with flexibility by using very weak learners as the trees, which makes it suitable for capturing complex functions without specifying a particular functional form beyond the link and without overfitting.

References

Hill, J., Weiss, C., & Zhai, F. (2011). Challenges With Propensity Score Strategies in a High-Dimensional Setting and a Potential Alternative. Multivariate Behavioral Research, 46(3), 477–513. doi: 10.1080/00273171.2011.570161

Chipman, H. A., George, E. I., & McCulloch, R. E. (2010). BART: Bayesian additive regression trees. The Annals of Applied Statistics, 4(1), 266–298. doi: 10.1214/09-AOAS285

Note that many references that deal with BART for causal inference focus on estimating potential outcomes with BART, not the propensity scores, and so are not directly relevant when using BART to estimate propensity scores for weights.

See method_ps for additional references.

See also

Examples

library("cobalt")
data("lalonde", package = "cobalt")

#Balancing covariates between treatment groups (binary)
(W1 <- weightit(treat ~ age + educ + married +
                  nodegree + re74, data = lalonde,
                method = "bart", estimand = "ATT",
                mc.cores = 4))

#> A weightit object
#>  - method: "bart" (propensity score weighting with BART)
#>  - number of obs.: 614
#>  - sampling weights: none
#>  - treatment: 2-category
#>  - estimand: ATT (focal: 1)
#>  - covariates: age, educ, married, nodegree, re74

summary(W1)

#>                  Summary of weights
#> 
#> - Weight ranges:
#> 
#>            Min                                  Max
#> treated 1.0000     ||                        1.0000
#> control 0.0075 |---------------------------| 7.0014
#> 
#> - Units with 5 greatest weights by group:
#>                                            
#>               1      2      3      4      6
#>  treated      1      1      1      1      1
#>             608    592    569    409    374
#>  control 2.1599 2.4093 2.4224 2.6899 7.0014
#> 
#> - Weight statistics:
#> 
#>         Coef of Var   MAD Entropy # Zeros
#> treated       0.000 0.000  -0.000       0
#> control       1.476 0.861  -0.461       0
#> overall       0.961 0.761  -0.205       0
#> 
#> - Effective Sample Sizes:
#> 
#>            Control Treated
#> Unweighted  429.       185
#> Weighted    135.24     185

bal.tab(W1)

#> Call
#>  weightit(formula = treat ~ age + educ + married + nodegree + 
#>     re74, data = lalonde, method = "bart", estimand = "ATT", 
#>     mc.cores = 4)
#> 
#> Balance Measures
#>                Type Diff.Adj
#> prop.score Distance   0.5153
#> age         Contin.   0.0561
#> educ        Contin.  -0.0429
#> married      Binary  -0.0367
#> nodegree     Binary   0.0422
#> re74        Contin.  -0.0859
#> 
#> Effective sample sizes
#>            Control Treated
#> Unadjusted  429.       185
#> Adjusted    135.24     185

#Balancing covariates with respect to race (multinomial)
(W2 <- weightit(race ~ age + educ + married +
                  nodegree + re74, data = lalonde,
                method = "bart", estimand = "ATE",
                mc.cores = 4))

#> A weightit object
#>  - method: "bart" (propensity score weighting with BART)
#>  - number of obs.: 614
#>  - sampling weights: none
#>  - treatment: 3-category (black, hispan, white)
#>  - estimand: ATE
#>  - covariates: age, educ, married, nodegree, re74

summary(W2)

#>                  Summary of weights
#> 
#> - Weight ranges:
#> 
#>           Min                                   Max
#> black  1.2182  |---------------|             9.0538
#> hispan 2.1677    |------------------------| 13.9344
#> white  1.0774 |----------------|             8.7787
#> 
#> - Units with 5 greatest weights by group:
#>                                                
#>             423     283     244     231     226
#>   black  7.3073   7.385  7.9828  8.3939  9.0538
#>             570     564     345     180      28
#>  hispan 12.4842 12.4969 13.0479 13.8225 13.9344
#>             409     140      76      60      23
#>   white  4.8468  5.4019  5.5876  8.6457  8.7787
#> 
#> - Weight statistics:
#> 
#>         Coef of Var   MAD Entropy # Zeros
#> black         0.604 0.393   0.972       0
#> hispan        0.415 0.345   2.056       0
#> white         0.489 0.325   0.756       0
#> overall       0.821 0.547   1.234       0
#> 
#> - Effective Sample Sizes:
#> 
#>             black hispan  white
#> Unweighted 243.    72.   299.  
#> Weighted   178.32  61.53 241.52

bal.tab(W2)

#> Call
#>  weightit(formula = race ~ age + educ + married + nodegree + re74, 
#>     data = lalonde, method = "bart", estimand = "ATE", mc.cores = 4)
#> 
#> Balance summary across all treatment pairs
#>             Type Max.Diff.Adj
#> age      Contin.       0.1961
#> educ     Contin.       0.1917
#> married   Binary       0.0423
#> nodegree  Binary       0.0491
#> re74     Contin.       0.1293
#> 
#> Effective sample sizes
#>             black hispan  white
#> Unadjusted 243.    72.   299.  
#> Adjusted   178.32  61.53 241.52

#Balancing covariates with respect to re75 (continuous)
#assuming t(3) conditional density for treatment
(W3 <- weightit(re75 ~ age + educ + married +
                  nodegree + re74, data = lalonde,
                method = "bart", density = "dt_3",
                mc.cores = 4))

#> A weightit object
#>  - method: "bart" (propensity score weighting with BART)
#>  - number of obs.: 614
#>  - sampling weights: none
#>  - treatment: continuous
#>  - covariates: age, educ, married, nodegree, re74

summary(W3)

#>                  Summary of weights
#> 
#> - Weight ranges:
#> 
#>        Min                                   Max
#> all 0.1353 |---------------------------| 18.0418
#> 
#> - Units with 5 greatest weights by group:
#>                                         
#>        490    487    484     469     310
#>  all 7.291 8.0573 8.4325 17.5039 18.0418
#> 
#> - Weight statistics:
#> 
#>     Coef of Var   MAD Entropy # Zeros
#> all       1.078 0.469   0.449       0
#> 
#> - Effective Sample Sizes:
#> 
#>             Total
#> Unweighted 614.  
#> Weighted   284.34

bal.tab(W3)

#> Call
#>  weightit(formula = re75 ~ age + educ + married + nodegree + re74, 
#>     data = lalonde, method = "bart", density = "dt_3", mc.cores = 4)
#> 
#> Balance Measures
#>             Type Corr.Adj
#> age      Contin.   0.0320
#> educ     Contin.   0.0525
#> married   Binary   0.0846
#> nodegree  Binary  -0.0798
#> re74     Contin.   0.1213
#> 
#> Effective sample sizes
#>             Total
#> Unadjusted 614.  
#> Adjusted   284.34