Entropy Balancing

This page explains the details of estimating weights using entropy balancing by setting method = "ebal" in the call to weightit or weightitMSM. This method can be used with binary, multinomial, and continuous treatments.

In general, this method relies on estimating weights by minimizing the entropy of the weights subject to exact moment balancing constraints. This method relies on ebalance from the ebal package for binary and multinomial treatments and on code written for WeightIt using optim for continuous treatments.

Binary Treatments

For binary treatments, this method estimates the weights using ebalance. The following estimands are allowed: ATE, ATT, and ATC. The weights are taken from the output of the ebalance fit object. When the ATE is requested, ebalance is run twice, once for each treatment group.

Multinomial Treatments

For multinomial treatments, this method estimates the weights using ebalance. The following estimands are allowed: ATE and ATT. The weights are taken from the output of the ebalance fit objects. When the ATE is requested, ebalance is run once for each treatment group. When the ATT is requested, ebalance is run once for each non-focal (i.e., control) group.

Continuous Treatments

For continuous treatments, this method estimates the weights using optim using formulas described by Hainmueller (2012) and Tübbicke (2020).

Longitudinal Treatments

For longitudinal treatments, the weights are the product of the weights estimated at each time point. This method is not guaranteed to yield exact balance at each time point. NOTE: the use of entropy balancing with longitudinal treatments has not been validated!

Sampling Weights

Sampling weights are supported through s.weights in all scenarios.

Missing Data

In the presence of missing data, the following value(s) for missing are allowed:

"ind" (default)

First, for each variable with missingness, a new missingness indicator variable is created which takes the value 1 if the original covariate is NA and 0 otherwise. The missingness indicators are added to the model formula as main effects. The missing values in the covariates are then replaced with 0s (this value is arbitrary and does not affect estimation). The weight estimation then proceeds with this new formula and set of covariates. The covariates output in the resulting weightit object will be the original covariates with the NAs.

Additional Arguments

All arguments to ebalance can be passed through weightit or weightitMSM and take on the defaults of those in ebalance.

For base.weights, a vector with length equal to the total number of units can be supplied, in contrast to ebalance(), which requires a vector with length equal to the number of controls. This works for continuous treatments as well. These correspond to the base weights q in Hainmueller (2012). The estimated weights minimize the Kullback entropy divergence from the base weights, defined as \(w*log(w/q)\), subject to exact balance constraints. These can be used to supply previously estimated weights so that the newly estimated weights retain the some of the proprties of the original weights while ensuring the balance constraints are met.

When standardize = TRUE in the call to weightit for binary and multinomial treatments, ebalance.trim is run on the resulting ebalance fit objects. Doing so can reduce the variability of the weights while maintaining covariate balance.

Additional Outputs

obj

When include.obj = TRUE, the entropy balancing model fit. For binary treatments with estimand = "ATT", the output of the call to ebal::ebalance or ebal::ebalance.trim when stabilize = TRUE. For binary treatments with estimand = "ATE" and multinomial treatments, a list of outputs of calls to ebal::ebalance or ebal::ebalance.trim. For continuous treatments, the output of the call to optim, which contains the dual variables and convergence information.

Details

Entropy balancing involves the specification of an optimization problem, the solution to which is then used to compute the weights. The constraints of the primal optimization problem correspond to covariate balance on the means (for binary and multinomial tretaments) or treatment-covariate covariances (for continuous treatments), positivity of the weights, and that the weights sum to a certain value. It turns out that the dual optimization problem is much easier to solve because it is over only as many variables as there are balance constraints rather than over the weights for each unit and it is unconstrained. Zhao and Percival (2017) found that entropy balancing for the ATT of a binary treatment actually involves the estimation of the coeffcients of a logistic regression propensity score model but using a specialized loss function different from that optimized with maximum likelihood. Entropy balancing is doubly robust (for the ATT) in the sense that it is consistent either when the true propensity score model is a logistic regression of the treatment on the covariates or when the trueoutcome model for the control units is a linear regression of the outcome on the covariates, and it attains a semi-parametric efficiency bound when both are true. Entropy balancing will always yield exact balance on the included terms.

Author

The code for continuous treatments was kindly contributed by Stefan Tübbicke (tuebbicke@empwifo.uni-potsdam.de).

References

- Binary Treatments

Hainmueller, J. (2012). Entropy Balancing for Causal Effects: A Multivariate Reweighting Method to Produce Balanced Samples in Observational Studies. Political Analysis, 20(1), 25–46. doi: 10.1093/pan/mpr025

Zhao, Q., & Percival, D. (2017). Entropy balancing is doubly robust. Journal of Causal Inference, 5(1). doi: 10.1515/jci-2016-0010

- Continuous Treatments

Tübbicke, S. (2020). Entropy Balancing for Continuous Treatments. ArXiv:2001.06281 [Econ, Stat]. http://arxiv.org/abs/2001.06281

See also

Examples

library("cobalt")
data("lalonde", package = "cobalt")

#Balancing covariates between treatment groups (binary)
(W1 <- weightit(treat ~ age + educ + married +
                  nodegree + re74, data = lalonde,
                method = "ebal", estimand = "ATT"))

#> A weightit object
#>  - method: "ebal" (entropy balancing)
#>  - number of obs.: 614
#>  - sampling weights: none
#>  - treatment: 2-category
#>  - estimand: ATT (focal: 1)
#>  - covariates: age, educ, married, nodegree, re74

summary(W1)

#>                  Summary of weights
#> 
#> - Weight ranges:
#> 
#>            Min                                  Max
#> treated 1.0000              ||               1.0000
#> control 0.0172 |---------------------------| 2.2624
#> 
#> - Units with 5 greatest weights by group:
#>                                            
#>               1      2      3      4      5
#>  treated      1      1      1      1      1
#>             595    589    409    296    269
#>  control 1.4643 1.4847 1.5763 1.7433 2.2624
#> 
#> - Weight statistics:
#> 
#>         Coef of Var   MAD Entropy # Zeros
#> treated       0.000 0.000    0.00       0
#> control       0.838 0.707   -0.50       0
#> overall       0.663 0.606   -0.25       0
#> 
#> - Effective Sample Sizes:
#> 
#>            Control Treated
#> Unweighted 429.000     185
#> Weighted   252.163     185

bal.tab(W1)

#> Call
#>  weightit(formula = treat ~ age + educ + married + nodegree + re74, data = lalonde, method = "ebal", estimand = "ATT")
#> 
#> Balance Measures
#>             Type Diff.Adj
#> age      Contin.  -0.0001
#> educ     Contin.   0.0000
#> married   Binary   0.0000
#> nodegree  Binary   0.0000
#> re74     Contin.  -0.0002
#> 
#> Effective sample sizes
#>            Control Treated
#> Unadjusted 429.000     185
#> Adjusted   252.163     185

#Balancing covariates with respect to race (multinomial)
(W2 <- weightit(race ~ age + educ + married +
                  nodegree + re74, data = lalonde,
                method = "ebal", estimand = "ATE",
                standardize = TRUE))

#> A weightit object
#>  - method: "ebal" (entropy balancing)
#>  - number of obs.: 614
#>  - sampling weights: none
#>  - treatment: 3-category (black, hispan, white)
#>  - estimand: ATE
#>  - covariates: age, educ, married, nodegree, re74

summary(W2)

#>                  Summary of weights
#> 
#> - Weight ranges:
#> 
#>           Min                                   Max
#> black  1.3975 |-----------|                 13.5172
#> hispan 1.2102 |---------------------------| 28.4321
#> white  0.8170 |-|                            3.9492
#> 
#> - Units with 5 greatest weights by group:
#>                                                
#>             485     244     226     182     181
#>   black  6.3711  6.4409    7.09  8.9826 13.5172
#>             564     392     371     345     269
#>  hispan 17.4558 21.5908 22.4375 23.0408 28.4321
#>             599     589     531     457      68
#>   white  3.5134  3.5373  3.5799   3.643  3.9492
#> 
#> - Weight statistics:
#> 
#>         Coef of Var   MAD Entropy # Zeros
#> black         0.590 0.413   1.058       0
#> hispan        0.609 0.439   2.307       0
#> white         0.371 0.306   0.787       0
#> overall       0.966 0.560   1.384       0
#> 
#> - Effective Sample Sizes:
#> 
#>              black hispan   white
#> Unweighted 243.000  72.00 299.000
#> Weighted   180.477  52.74 262.932

bal.tab(W2)

#> Call
#>  weightit(formula = race ~ age + educ + married + nodegree + re74, data = lalonde, method = "ebal", estimand = "ATE", standardize = TRUE)
#> 
#> Balance summary across all treatment pairs
#>             Type Max.Diff.Adj
#> age      Contin.       0.0005
#> educ     Contin.       0.0012
#> married   Binary       0.0002
#> nodegree  Binary       0.0002
#> re74     Contin.       0.0001
#> 
#> Effective sample sizes
#>              black hispan   white
#> Unadjusted 243.000  72.00 299.000
#> Adjusted   180.477  52.74 262.932

#Balancing covariates with respect to re75 (continuous)
(W3 <- weightit(re75 ~ age + educ + married +
                  nodegree + re74, data = lalonde,
                method = "ebal"))

#> A weightit object
#>  - method: "ebal" (entropy balancing)
#>  - number of obs.: 614
#>  - sampling weights: none
#>  - treatment: continuous
#>  - covariates: age, educ, married, nodegree, re74

summary(W3)

#>                  Summary of weights
#> 
#> - Weight ranges:
#> 
#>     Min                                   Max
#> all   0 |---------------------------| 10.5998
#> 
#> - Units with 5 greatest weights by group:
#>                                         
#>         483    482    481    180     178
#>  all 4.5868 4.7046 4.8506 7.9044 10.5998
#> 
#> - Weight statistics:
#> 
#>     Coef of Var   MAD Entropy # Zeros
#> all       0.765 0.434   0.201       0
#> 
#> - Effective Sample Sizes:
#> 
#>              Total
#> Unweighted 614.000
#> Weighted   387.412

bal.tab(W3)

#> Call
#>  weightit(formula = re75 ~ age + educ + married + nodegree + re74, data = lalonde, method = "ebal")
#> 
#> Balance Measures
#>             Type Corr.Adj
#> age      Contin.       -0
#> educ     Contin.        0
#> married   Binary       -0
#> nodegree  Binary       -0
#> re74     Contin.       -0
#> 
#> Effective sample sizes
#>              Total
#> Unadjusted 614.000
#> Adjusted   387.412