Empirical Balancing Calibration Weighting

This page explains the details of estimating weights using empirical balancing calibration weighting (EBCW) by setting method = "ebcw" in the call to weightit. This method can be used with binary and multinomial treatments.

In general, this method relies on estimating weights by minimizing a function of the weights subject to exact moment balancing constraints. This method relies on ATE from the ATE package.

Binary Treatments

For binary treatments, this method estimates the weights using ATE with ATT = TRUE. The following estimands are allowed: ATE, ATT, and ATC. The weights are taken from the output of the ATE fit object. When the ATE is requested, ATE is run twice, once for each treatment group.

Multinomial Treatments

For multinomial treatments, this method estimates the weights using ATE with ATT = TRUE. The following estimands are allowed: ATE and ATT. The weights are taken from the output of the ATE fit objects. When the ATE is requested, ATE is run once for each treatment group. When the ATT is requested, ATE is run once for each non-focal (i.e., control) group.

Continuous Treatments

Continuous treatments are not supported.

Longitudinal Treatments

For longitudinal treatments, the weights are the product of the weights estimated at each time point. This method is not guaranteed to yield exact balance at each time point. NOTE: the use of EBCW with longitudinal treatments has not been validated!

Sampling Weights

Sampling weights are supported through s.weights in all scenarios.

Missing Data

In the presence of missing data, the following value(s) for missing are allowed:

"ind" (default)

First, for each variable with missingness, a new missingness indicator variable is created which takes the value 1 if the original covariate is NA and 0 otherwise. The missingness indicators are added to the model formula as main effects. The missing values in the covariates are then replaced with 0s (this value is arbitrary and does not affect estimation). The weight estimation then proceeds with this new formula and set of covariates. The covariates output in the resulting weightit object will be the original covariates with the NAs.

Additional Arguments

All argument to ATE can be passed through weightit, with the following exceptions:

ATT is ignored because the estimand is passed using estimand.

All arguments take on the defaults of those in ATE.

Additional Outputs

obj

When include.obj = TRUE, the empirical balancing callibration model fit. For binary treatments with estimand = "ATT", the output of the call to ATE::ATE. For binary treatments with estimand = "ATE" and multinomial treatments, a list of outputs of calls to ATE::ATE.

References

Chan, K. C. G., Yam, S. C. P., & Zhang, Z. (2016). Globally efficient non-parametric inference of average treatment effects by empirical balancing calibration weighting. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 78(3), 673–700. doi: 10.1111/rssb.12129

See also

Examples

library("cobalt")
data("lalonde", package = "cobalt")

#Balancing covariates between treatment groups (binary)
(W1 <- weightit(treat ~ age + educ + married +
                  nodegree + re74, data = lalonde,
                method = "ebcw", estimand = "ATT"))

#> A weightit object
#>  - method: "ebcw" (empirical balancing calibration weighting)
#>  - number of obs.: 614
#>  - sampling weights: none
#>  - treatment: 2-category
#>  - estimand: ATT (focal: 1)
#>  - covariates: age, educ, married, nodegree, re74

summary(W1)

#>                  Summary of weights
#> 
#> - Weight ranges:
#> 
#>            Min                                    Max
#> treated 1.0000                              || 1.0000
#> control 0.0001   ||                            0.0122
#> 
#> - Units with 5 greatest weights by group:
#>                                           
#>               1     2      3      4      5
#>  treated      1     1      1      1      1
#>             595   589    409    296    269
#>  control 0.0079 0.008 0.0085 0.0094 0.0122
#> 
#> - Weight statistics:
#> 
#>         Coef of Var   MAD Entropy # Zeros
#> treated       0.000 0.000   0.000       0
#> control       0.839 0.707  -5.721       0
#> overall       1.512 1.387  -0.031       0
#> 
#> - Effective Sample Sizes:
#> 
#>            Control Treated
#> Unweighted  429.       185
#> Weighted    252.12     185

bal.tab(W1)

#> Call
#>  weightit(formula = treat ~ age + educ + married + nodegree + re74, data = lalonde, method = "ebcw", estimand = "ATT")
#> 
#> Balance Measures
#>             Type Diff.Adj
#> age      Contin.       -0
#> educ     Contin.        0
#> married   Binary       -0
#> nodegree  Binary        0
#> re74     Contin.       -0
#> 
#> Effective sample sizes
#>            Control Treated
#> Unadjusted 429.000     185
#> Adjusted   252.121     185

#Balancing covariates with respect to race (multinomial)
(W2 <- weightit(race ~ age + educ + married +
                  nodegree + re74, data = lalonde,
                method = "ebcw", estimand = "ATE"))

#> A weightit object
#>  - method: "ebcw" (empirical balancing calibration weighting)
#>  - number of obs.: 614
#>  - sampling weights: none
#>  - treatment: 3-category (black, hispan, white)
#>  - estimand: ATE
#>  - covariates: age, educ, married, nodegree, re74

summary(W2)

#>                  Summary of weights
#> 
#> - Weight ranges:
#> 
#>           Min                                  Max
#> black  0.0023 |-----------|                 0.0220
#> hispan 0.0020 |---------------------------| 0.0463
#> white  0.0013 |-|                           0.0064
#> 
#> - Units with 5 greatest weights by group:
#>                                           
#>            485    244    226    182    181
#>   black 0.0104 0.0105 0.0115 0.0146  0.022
#>            564    392    371    345    269
#>  hispan 0.0284 0.0351 0.0366 0.0376 0.0463
#>            599    589    531    457     68
#>   white 0.0057 0.0058 0.0058 0.0059 0.0064
#> 
#> - Weight statistics:
#> 
#>         Coef of Var   MAD Entropy # Zeros
#> black         0.590 0.413  -5.362       0
#> hispan        0.609 0.440  -4.113       0
#> white         0.371 0.306  -5.633       0
#> overall       0.966 0.559  -5.036       0
#> 
#> - Effective Sample Sizes:
#> 
#>             black hispan  white
#> Unweighted 243.    72.   299.  
#> Weighted   180.47  52.71 262.93

bal.tab(W2)

#> Call
#>  weightit(formula = race ~ age + educ + married + nodegree + re74, data = lalonde, method = "ebcw", estimand = "ATE")
#> 
#> Balance summary across all treatment pairs
#>             Type Max.Diff.Adj
#> age      Contin.            0
#> educ     Contin.            0
#> married   Binary            0
#> nodegree  Binary            0
#> re74     Contin.            0
#> 
#> Effective sample sizes
#>              black hispan   white
#> Unadjusted 243.000 72.000 299.000
#> Adjusted   180.469 52.706 262.929