Optimization-Based Weighting

This page explains the details of estimating optimization-based weights 9also known as stable balancing weights) by setting method = "optweight" in the call to weightit or weightitMSM. This method can be used with binary, multinomial, and continuous treatments.

In general, this method relies on estimating weights by solving a quadratic programming problem subject to approximate or exact balance constraints. This method relies on optweight from the optweight package.

Because optweight offers finer control and uses the same syntax as weightit, it is recommended that optweight be used instead of weightit with method = "optweight".

Binary Treatments

For binary treatments, this method estimates the weights using optweight. The following estimands are allowed: ATE, ATT, and ATC. The weights are taken from the output of the optweight fit object.

Multinomial Treatments

For multinomial treatments, this method estimates the weights using optweight. The following estimands are allowed: ATE and ATT. The weights are taken from the output of the optweight fit object.

Continuous Treatments

For binary treatments, this method estimates the weights using optweight. The weights are taken from the output of the optweight fit object.

Longitudinal Treatments

For longitudinal treatments, optweight estimates weights that simultaneously satisfy balance constraints at all time points, so only one model is fit to obtain the weights. Using method = "optweight" in weightitMSM cause is.MSM.method to be set to TRUE by default. Setting it to FALSE will run one model for each time point and multiply the weights together, a method that is not recommended. NOTE: neither use of optimization-based weights with longitudinal treatments has been validated!

Sampling Weights

Sampling weights are supported through s.weights in all scenarios.

Missing Data

In the presence of missing data, the following value(s) for missing are allowed:

"ind" (default)

First, for each variable with missingness, a new missingness indicator variable is created which takes the value 1 if the original covariate is NA and 0 otherwise. The missingness indicators are added to the model formula as main effects. The missing values in the covariates are then replaced with 0s (this value is arbitrary and does not affect estimation). The weight estimation then proceeds with this new formula and set of covariates. The covariates output in the resulting weightit object will be the original covariates with the NAs.

Additional Arguments

All arguments to optweight can be passed through weightit or weightitMSM, with the following exception:

targets cannot be used and is ignored.

All arguments take on the defaults of those in optweight.

Additional Outputs

info

A list with one entry:

duals

A data frame of dual variables for each balance constraint.

obj

When include.obj = TRUE, the output of the call to optweight::optweight.

Note

The specification of tols differs between weightit and optweight. In weightit, one tolerance value should be included per level of each factor variable, whereas in optweight, all levels of a factor are given the same tolerance, and only one value needs to be supplied for a factor variable. Because of the potential for confusion and ambiguity, it is recommended to only supply one value for tols in weightit that applies to all variables. For finer control, use optweight directly.

Seriously, just use optweight. The syntax is almost identical and it's compatible with cobalt, too.

Details

Stable balancing weights are weights that solve a constrained optimization problem, where the constraints correspond to covariate balance and the loss function is the variance (or other norm) of the weights. These weights maximize the effective sample size of the weighted sample subject to user-supplied balance constraints. An advantage of this method over entropy balancing is the ability to allow approximate, rather than exact, balance through the tols argument, which can increase precision even for slight relaxations of the constraints.

References

Binary Treatments

Wang, Y., & Zubizarreta, J. R. (2020). Minimal dispersion approximately balancing weights: Asymptotic properties and practical considerations. Biometrika, 107(1), 93–105. doi: 10.1093/biomet/asz050

Zubizarreta, J. R. (2015). Stable Weights that Balance Covariates for Estimation With Incomplete Outcome Data. Journal of the American Statistical Association, 110(511), 910–922. doi: 10.1080/01621459.2015.1023805

Multinomial Treatments

de los Angeles Resa, M., & Zubizarreta, J. R. (2020). Direct and stable weight adjustment in non-experimental studies with multivalued treatments: Analysis of the effect of an earthquake on post-traumatic stress. Journal of the Royal Statistical Society: Series A (Statistics in Society), n/a(n/a). doi: 10.1111/rssa.12561

Continuous Treatments

Greifer, N. (2020). Estimating Balancing Weights for Continuous Treatments Using Constrained Optimization. doi: 10.17615/DYSS-B342

See also

Examples

library("cobalt")
data("lalonde", package = "cobalt")

#Balancing covariates between treatment groups (binary)
(W1 <- weightit(treat ~ age + educ + married +
                  nodegree + re74, data = lalonde,
                method = "optweight", estimand = "ATT",
                tols = 0))

#> A weightit object
#>  - method: "optweight" (targeted stable balancing weights)
#>  - number of obs.: 614
#>  - sampling weights: none
#>  - treatment: 2-category
#>  - estimand: ATT (focal: 1)
#>  - covariates: age, educ, married, nodegree, re74

summary(W1)

#>                  Summary of weights
#> 
#> - Weight ranges:
#> 
#>         Min                                  Max
#> treated   1           ||                  1.0000
#> control   0 |---------------------------| 3.0426
#> 
#> - Units with 5 greatest weights by group:
#>                                            
#>               1      2      3      4      5
#>  treated      1      1      1      1      1
#>             589    411    409    296    269
#>  control 2.5261 2.5415 2.6434 2.7396 3.0426
#> 
#> - Weight statistics:
#> 
#>         Coef of Var   MAD Entropy # Zeros
#> treated       0.000 0.000  -0.000       0
#> control       0.788 0.697   0.393      83
#> overall       0.658 0.487   0.275      83
#> 
#> - Effective Sample Sizes:
#> 
#>            Control Treated
#> Unweighted  429.       185
#> Weighted    264.88     185

bal.tab(W1)

#> Call
#>  weightit(formula = treat ~ age + educ + married + nodegree + 
#>     re74, data = lalonde, method = "optweight", estimand = "ATT", 
#>     tols = 0)
#> 
#> Balance Measures
#>             Type Diff.Adj
#> age      Contin.       -0
#> educ     Contin.       -0
#> married   Binary       -0
#> nodegree  Binary        0
#> re74     Contin.       -0
#> 
#> Effective sample sizes
#>            Control Treated
#> Unadjusted  429.       185
#> Adjusted    264.88     185

#Balancing covariates with respect to race (multinomial)
(W2 <- weightit(race ~ age + educ + married +
                  nodegree + re74, data = lalonde,
                method = "optweight", estimand = "ATE",
                tols = .01))

#> A weightit object
#>  - method: "optweight" (targeted stable balancing weights)
#>  - number of obs.: 614
#>  - sampling weights: none
#>  - treatment: 3-category (black, hispan, white)
#>  - estimand: ATE
#>  - covariates: age, educ, married, nodegree, re74

summary(W2)

#>                  Summary of weights
#> 
#> - Weight ranges:
#> 
#>           Min                                  Max
#> black  0.4429     |-----------------------| 3.5741
#> hispan 0.0000 |-------------------|         2.5848
#> white  0.2574   |---------|                 1.6593
#> 
#> - Units with 5 greatest weights by group:
#>                                           
#>            485    190    184    182    181
#>   black 2.3351 2.3723 2.5586 2.8367 3.5741
#>            564    392    371    345    269
#>  hispan 2.0459 2.0984 2.1887 2.1982 2.5848
#>            599    589    531    324     68
#>   white 1.5706 1.5706 1.5725 1.5968 1.6593
#> 
#> - Weight statistics:
#> 
#>         Coef of Var   MAD Entropy # Zeros
#> black         0.550 0.443   0.130       0
#> hispan        0.566 0.449   0.176       2
#> white         0.353 0.295   0.065       0
#> overall       0.466 0.372   0.104       2
#> 
#> - Effective Sample Sizes:
#> 
#>             black hispan  white
#> Unweighted 243.     72.  299.  
#> Weighted   186.76   54.7 266.01

bal.tab(W2)

#> Call
#>  weightit(formula = race ~ age + educ + married + nodegree + re74, 
#>     data = lalonde, method = "optweight", estimand = "ATE", tols = 0.01)
#> 
#> Balance summary across all treatment pairs
#>             Type Max.Diff.Adj
#> age      Contin.         0.01
#> educ     Contin.         0.01
#> married   Binary         0.01
#> nodegree  Binary         0.01
#> re74     Contin.         0.01
#> 
#> Effective sample sizes
#>             black hispan  white
#> Unadjusted 243.     72.  299.  
#> Adjusted   186.76   54.7 266.01

#Balancing covariates with respect to re75 (continuous)
(W3 <- weightit(re75 ~ age + educ + married +
                  nodegree + re74, data = lalonde,
                method = "optweight", tols = .05))

#> A weightit object
#>  - method: "optweight" (targeted stable balancing weights)
#>  - number of obs.: 614
#>  - sampling weights: none
#>  - treatment: continuous
#>  - covariates: age, educ, married, nodegree, re74

summary(W3)

#>                  Summary of weights
#> 
#> - Weight ranges:
#> 
#>     Min                                 Max
#> all   0 |---------------------------| 3.966
#> 
#> - Units with 5 greatest weights by group:
#>                                      
#>         483    482   200    180   178
#>  all 2.9566 3.0117 3.048 3.5934 3.966
#> 
#> - Weight statistics:
#> 
#>     Coef of Var   MAD Entropy # Zeros
#> all       0.574 0.445   0.173      34
#> 
#> - Effective Sample Sizes:
#> 
#>             Total
#> Unweighted 614.  
#> Weighted   462.02

bal.tab(W3)

#> Call
#>  weightit(formula = re75 ~ age + educ + married + nodegree + re74, 
#>     data = lalonde, method = "optweight", tols = 0.05)
#> 
#> Balance Measures
#>             Type Corr.Adj
#> age      Contin.  -0.0087
#> educ     Contin.  -0.0106
#> married   Binary   0.0499
#> nodegree  Binary  -0.0451
#> re74     Contin.   0.0499
#> 
#> Effective sample sizes
#>             Total
#> Unadjusted 614.  
#> Adjusted   462.02