Print and Summarize Output

summary() generates a summary of the weightit or weightitMSM object to evaluate the properties of the estimated weights. plot() plots the distribution of the weights. nobs() extracts the number of observations.

Usage

# S3 method for class 'weightit'
summary(object, top = 5L, ignore.s.weights = FALSE, ...)

# S3 method for class 'summary.weightit'
plot(x, binwidth = NULL, bins = NULL, ...)

# S3 method for class 'weightitMSM'
summary(object, top = 5L, ignore.s.weights = FALSE, ...)

# S3 method for class 'summary.weightitMSM'
plot(x, binwidth = NULL, bins = NULL, time = 1, ...)

Arguments

object

a weightit or weightitMSM object; the output of a call to weightit() or weightitMSM().

top

how many of the largest and smallest weights to display. Default is 5.

ignore.s.weights

whether or not to ignore sampling weights when computing the weight summary. If FALSE, the default, the estimated weights will be multiplied by the sampling weights (if any) before values are computed.

...

For plot(), additional arguments passed to graphics::hist() to determine the number of bins, though ggplot2::geom_histogram() is actually used to create the plot.

x

a summary.weightit or summary.weightitMSM object; the output of a call to summary.weightit() or summary.weightitMSM().

binwidth, bins

arguments passed to ggplot2::geom_histogram() to control the size and/or number of bins.

time

numeric; the time point for which to display the distribution of weights. Default is to plot the distribution for the first time points.

Value

For point treatments (i.e., weightit objects), summary() returns a summary.weightit object with the following elements:

weight.range

The range (minimum and maximum) weight for each treatment group.

weight.top

The units with the greatest weights in each treatment group; how many are included is determined by top.

coef.of.var (Coef of Var)

The coefficient of variation (standard deviation divided by mean) of the weights in each treatment group and overall.

scaled.mad (MAD)

The mean absolute deviation of the weights in each treatment group and overall divided by the mean of the weights in the corresponding group.

negative entropy (Entropy)

The negative entropy (\(\sum w log(w)\)) of the weights in each treatment group and overall divided by the mean of the weights in the corresponding group.

num.zeros

The number of weights equal to zero.

effective.sample.size

The effective sample size for each treatment group before and after weighting. See ESS().

For longitudinal treatments (i.e., weightitMSM objects), summary() returns a list of the above elements for each treatment period.

plot() returns a ggplot object with a histogram displaying the distribution of the estimated weights. If the estimand is the ATT or ATC, only the weights for the non-focal group(s) will be displayed (since the weights for the focal group are all 1). A dotted line is displayed at the mean of the weights.

nobs() returns a single number. Note that even units with weights or s.weights of 0 are included.

See also

Examples

# See example at ?weightit or ?weightitMSM